+++ ChangeLog.md
@@ -6,6 +6,7 @@ https://bitbucket.org/art-uniroma2/semantic-turkey/src/master/ChangeLog.txt
 
 # 7.0.1 (dd-mm-yyyy)
   * Enabled organization of projects in directories
+  * Enabled the AND-scheme for browsing concept tree
 
 # 7.0.0 (10-02-2020)
   * Introduced EDOAL projects.